Fixing Factory girl.
Factory girl couldn't be read in the support file. Then Database cleaner didn't exist so we have to grab that gem and bundle it in. Meanwhile the rails_helper file is looking kinda ugly so I may clean that up next. I first tried to troubleshoot the factory girl by doublechecking the factories and let blocks calling the methods. Those all checked out after I thoroughly changed them all. I then reverted back and began looking into the gem being the issue. It wasn't, but I noticed the support file wasn't being read by Rspec. Rspec is weak in this degree and the answer was not posted on Thoughtbot's readme for the gem. I simply moved the support file's contents to the rails helper and that gave me a different error. The test still ran, but DatabaseCleaner wasn't recognized. On commenting that out, the spec works and gives me errors on not having the models.