Completed another guest login test.
The log in button works, but we are having an issue trying to stub session. Getting session to work is easy, testing it is hard. I figure cart will be the same. We may have to stray from tdd as a workaround due to time constraints. For now we will simply not test the current user in hopes we can get help about it tomorrow.