Finished final spec for guest login.
I used a flash which led me to add something to view layout. That block of code will display the flash in the same manner across all views and it also includes a tag for easy testing and frontending. I actually may work on the authorization earlier than expected. Looking through old code I'm reminded how easy it is.