Finished first Admin dashboard user test.
I added capybara, launchy, and cancancan to the gemfile. I am beginning working on authorization as part of the admin dashboard. I realized the escape will be easy because it won't leave a message and will always redirect to the user dashboard. Also this enum method is very useful, I think we should switch the status on breeds and cats to it as well.