Cart is functional and clean.
We'll have to move a lot of the logic out of the controller, but I think that can easily be solved with a CurrentCart class. We're passing the cat.id twice so we can redirect to the previous page. Perhaps the dashboard would be easier and the way to go.