Seed data updated.
In this commit we created a migration to change the default of cat.image_path. We then added logic to the cat breed to set the default cat image to the breed image if none is initialized. We also added validations to the relationships of breed and cats. The main goal and effect of this commit is to create the seed data with random cool stuff. We implemented it mostly, but still have to tweak the names and un-lorem ipsum the descriptions.